+++ Projektovanje/backend/efikas/src/main/java/org/unibl/etf/efikas/repositories/ApartmentRepository.java
@@ -7,8 +7,8 @@ import org.unibl.etf.efikas.models.entities.Apartment;
 import java.util.List;
 
 @Repository
-public interface ApartmentRepository extends JpaRepository<Apartment, Long> {
-    Apartment findApartmentByApartmentId(Long apartmentId);
+public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {
+    Apartment findApartmentByApartmentId(Integer apartmentId);
     Apartment findApartmentByAddress(String apartmentAddress);
     List<Apartment> findByUserEmail(String email);
 }
